requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.4
django-filter==24.2
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
phonenumberslite==8.13.36
pillow==10.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.0
sqlparse==0.5.0
uritemplate==4.1.1
